Rename the socket.
